Add PlayerState
The position of each player was previously stored as an integer (which
was always changing). Having an immutable object to represent player
state is not only cleaner, but will prevent any unwanted changes from
being made to a player's state.

Usage wise, another object will house a collection of player states to
keep track of the game's progression.