Add additional methods to game
In order to print out a Game, there was extra bits of information that
were needed by the ConsolePrinter. The functionality needed isn't
specific to printing out the game, so it was added directly to the Game
class.

For example, it's perfectly reasonable for a Game to be able to tell you
how many players are on a certain square, or give you a list of all the
players who are participating in the game.