Add last roll to player state
Previously, there was no way to know what numbers a player rolled other
than to keep track of the inputs that were given when moving a player.
With this change, we can not only know what a player just rolled, but
also all of the rolls that have happened in a game.