Modify API of Game
While testing Game, it became clear that the method for determining if
the last move was a win was somewhat brittle. For example, if the check
wasn't made immediately after the player made the winning move, there
wouldn't be a way to check if that player had already won the game.